+++ pkg/plugins/autodiscovery/kubernetes/containers.go
@@ -16,185 +16,6 @@ var (
 	DefaultKubernetesFiles []string = []string{"*.yaml", "*.yml"}
 )
 
-func (k Kubernetes) discoverKubernetesManifest(file, relativeFile string) [][]byte {
-	var manifests [][]byte
-
-	data, err := getKubernetesManifestData(file)
-	if err != nil {
-		logrus.Debugln(err)
-		return manifests
-	}
-	if data == nil {
-		return manifests
-	}
-	for i, container := range data.Spec.Containers {
-
-		containerName := container.Name
-		if containerName == "" {
-			containerName = container.Image
-		}
-
-		manifest, err := k.generateContainerManifest(
-			fmt.Sprintf("$.spec.containers[%d].image", i),
-			containerName,
-			container.Image,
-			relativeFile,
-			"")
-
-		if err != nil {
-			logrus.Debugln(err)
-			continue
-		}
-
-		if manifest == nil {
-			continue
-		}
-
-		manifests = append(manifests, manifest)
-	}
-
-	for i, container := range data.Spec.Template.Spec.Containers {
-
-		containerName := container.Name
-		if containerName == "" {
-			containerName = container.Image
-		}
-
-		manifest, err := k.generateContainerManifest(
-			fmt.Sprintf("$.spec.template.spec.containers[%d].image", i),
-			containerName,
-			container.Image,
-			relativeFile,
-			"")
-
-		if err != nil {
-			logrus.Debugln(err)
-			continue
-		}
-
-		if manifest == nil {
-			continue
-		}
-
-		manifests = append(manifests, manifest)
-	}
-	return manifests
-}
-
-func (k Kubernetes) discoverProwManifest(file, relativeFile string) [][]byte {
-	var manifests [][]byte
-
-	data, err := getProwManifestData(file)
-	if err != nil {
-		logrus.Debugln(err)
-		return manifests
-	}
-	if data == nil {
-		return manifests
-	}
-	// Prow Presubmit
-	for repo, tests := range data.ProwPreSubmitJobs {
-		for i, test := range tests {
-			for j, container := range test.Spec.Containers {
-				containerName := container.Name
-				if containerName == "" {
-					imageName, _, _, err := dockerimage.ParseOCIReferenceInfo(container.Image)
-					if err != nil {
-						logrus.Debugln(err)
-						continue
-					}
-					containerName = imageName
-				}
-				manifest, err := k.generateContainerManifest(
-					fmt.Sprintf("$.presubmits.'%s'[%d].spec.containers[%d].image", repo, i, j),
-					containerName,
-					container.Image,
-					relativeFile,
-					fmt.Sprintf(" for repo %q and presubmit test %q", repo, test.Name))
-
-				if err != nil {
-					logrus.Debugln(err)
-					continue
-				}
-
-				if manifest == nil {
-					continue
-				}
-
-				manifests = append(manifests, manifest)
-			}
-		}
-	}
-
-	// Prow Postsubmit
-	for repo, tests := range data.ProwPostSubmitJobs {
-		for i, test := range tests {
-			for j, container := range test.Spec.Containers {
-				containerName := container.Name
-				if containerName == "" {
-					imageName, _, _, err := dockerimage.ParseOCIReferenceInfo(container.Image)
-					if err != nil {
-						logrus.Debugln(err)
-						continue
-					}
-					containerName = imageName
-				}
-				manifest, err := k.generateContainerManifest(
-					fmt.Sprintf("$.postsubmits.'%s'[%d].spec.containers[%d].image", repo, i, j),
-					containerName,
-					container.Image,
-					relativeFile,
-					fmt.Sprintf(" for repo %q and postsubmit test %q", repo, test.Name))
-
-				if err != nil {
-					logrus.Debugln(err)
-					continue
-				}
-
-				if manifest == nil {
-					continue
-				}
-
-				manifests = append(manifests, manifest)
-			}
-		}
-	}
-
-	// Prow Periodics
-	for i, test := range data.ProwPeriodicJobs {
-		for j, container := range test.Spec.Containers {
-			containerName := container.Name
-			if containerName == "" {
-
-				imageName, _, _, err := dockerimage.ParseOCIReferenceInfo(container.Image)
-				if err != nil {
-					logrus.Debugln(err)
-					continue
-				}
-				containerName = imageName
-			}
-			manifest, err := k.generateContainerManifest(
-				fmt.Sprintf("$.periodics[%d].spec.containers[%d].image", i, j),
-				containerName,
-				container.Image,
-				relativeFile,
-				fmt.Sprintf(" for periodic test %q", test.Name))
-
-			if err != nil {
-				logrus.Debugln(err)
-				continue
-			}
-
-			if manifest == nil {
-				continue
-			}
-
-			manifests = append(manifests, manifest)
-		}
-	}
-	return manifests
-}
-
 func (k Kubernetes) discoverContainerManifests() ([][]byte, error) {
 	var manifests [][]byte
 	searchFromDir := k.rootDir
@@ -217,17 +38,18 @@ func (k Kubernetes) discoverContainerManifests() ([][]byte, error) {
 
 		relativeFoundKubernetesFile, err := filepath.Rel(k.rootDir, kubernetesFile)
 		if err != nil {
-			// Let's try the next chart if one fail
 			logrus.Debugln(err)
 			continue
 		}
 
-		if k.flavor == FlavorKubernetes {
+		switch k.flavor {
+		case FlavorKubernetes:
 			manifests = append(manifests, k.discoverKubernetesManifest(kubernetesFile, relativeFoundKubernetesFile)...)
-		} else if k.flavor == FlavorProw {
+		case FlavorProw:
 			manifests = append(manifests, k.discoverProwManifest(kubernetesFile, relativeFoundKubernetesFile)...)
+		default:
+			return nil, fmt.Errorf("Kubernetes manifest %q not supported", k.flavor)
 		}
-
 	}
 
 	return manifests, nil
